move resolver to separate file
